Write shortlist_rank as a nullable integer
The 3,170 unranked localities need blanks, and a plain int column cannot hold
them, so pandas silently widened the column to float and wrote "1.0" into the
CSV. Power BI then read it as a decimal and displayed ranks as 1.0, 2.0, 10.0
in the report table, with no way to fix it from the report side.

Using pandas' nullable Int64 keeps the blanks and writes 1, 2, 10.

+++ src/build_dashboard_dataset.py
@@ -140,9 +140,14 @@ def build_serving_table(df: pd.DataFrame) -> pd.DataFrame:
 
     shortlist_index = select_shortlist(df)
     out["is_shortlisted"] = out.index.isin(shortlist_index)
-    out["shortlist_rank"] = pd.Series(
-        range(1, len(shortlist_index) + 1), index=shortlist_index
-    ).reindex(out.index)
+    # Nullable integer, not float. A plain int column cannot hold the blanks the
+    # 3,170 unranked localities need, so pandas would silently widen it to float and
+    # write "1.0" into the CSV — which any BI tool then reads as a decimal.
+    out["shortlist_rank"] = (
+        pd.Series(range(1, len(shortlist_index) + 1), index=shortlist_index)
+        .reindex(out.index)
+        .astype("Int64")
+    )
 
     return out
